Conflicts:
	Der grosse
Dalmuti/src/ch/fhnw/error404/DerGrosseDalmuti/client/Client_neu.java

+++ Der grosse Dalmuti/src/ch/fhnw/error404/DerGrosseDalmuti/client/Client_neu.java
@@ -15,7 +15,7 @@ import java.net.Socket;
  */
 
 public class Client_neu {
-	Integer test = new Integer(689);
+	Integer test = new Integer(568);
 	ObjectInputStream in;
 
 	public static void main(String[] args) {
